Make sure routes are able to react to host updates.
