+++ pkg/manager/chemanager_controller.go
@@ -16,35 +16,21 @@ import (
 	"context"
 
 	"github.com/che-incubator/devworkspace-che-operator/apis/che-controller/v1alpha1"
-	"github.com/che-incubator/devworkspace-che-operator/pkg/defaults"
 	"github.com/che-incubator/devworkspace-che-operator/pkg/infrastructure"
 	"github.com/che-incubator/devworkspace-che-operator/pkg/sync"
-	"github.com/google/go-cmp/cmp"
-	"github.com/google/go-cmp/cmp/cmpopts"
 	routev1 "github.com/openshift/api/route/v1"
 	appsv1 "k8s.io/api/apps/v1"
 	corev1 "k8s.io/api/core/v1"
 	"k8s.io/api/extensions/v1beta1"
 	rbac "k8s.io/api/rbac/v1"
 	"k8s.io/apimachinery/pkg/api/errors"
-	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
 	"k8s.io/apimachinery/pkg/runtime"
-	"k8s.io/apimachinery/pkg/util/intstr"
 	ctrl "sigs.k8s.io/controller-runtime"
 	"sigs.k8s.io/controller-runtime/pkg/client"
 )
 
 var (
 	log = ctrl.Log.WithName("che")
-
-	routeDiffOpts = cmp.Options{
-		cmpopts.IgnoreFields(routev1.Route{}, "TypeMeta", "ObjectMeta", "Status"),
-		cmpopts.IgnoreFields(routev1.RouteSpec{}, "WildcardPolicy"),
-	}
-
-	ingressDiffOpts = cmp.Options{
-		cmpopts.IgnoreFields(v1beta1.Ingress{}, "TypeMeta", "ObjectMeta", "Status"),
-	}
 )
 
 type CheReconciler struct {
@@ -152,94 +138,4 @@ func (r *CheReconciler) reconcileGateway(ctx context.Context, manager *v1alpha1.
 	}
 
 	return changed, err
-}
-
-func (r *CheReconciler) reconcileRoute(ctx context.Context, manager *v1alpha1.CheManager) (bool, error) {
-	route := getRouteSpec(manager)
-	var changed bool
-	var err error
-
-	if manager.Spec.Routing == v1alpha1.SingleHost {
-		changed, err = r.syncer.Sync(ctx, manager, route, routeDiffOpts)
-	} else {
-		changed, err = true, r.syncer.Delete(ctx, route)
-	}
-
-	return changed, err
-}
-
-func (r *CheReconciler) reconcileIngress(ctx context.Context, manager *v1alpha1.CheManager) (bool, error) {
-	ingress := getIngressSpec(manager)
-	var changed bool
-	var err error
-
-	if manager.Spec.Routing == v1alpha1.SingleHost {
-		changed, err = r.syncer.Sync(ctx, manager, ingress, ingressDiffOpts)
-	} else {
-		changed, err = true, r.syncer.Delete(ctx, ingress)
-	}
-
-	return changed, err
-}
-
-func getRouteSpec(manager *v1alpha1.CheManager) *routev1.Route {
-	return &routev1.Route{
-		ObjectMeta: metav1.ObjectMeta{
-			Name:      manager.Name,
-			Namespace: manager.Namespace,
-			Labels:    defaults.GetLabelsForComponent(manager, "external-access"),
-		},
-		Spec: routev1.RouteSpec{
-			Host: manager.Spec.Host,
-			To: routev1.RouteTargetReference{
-				Kind: "Service",
-				Name: GetGatewayServiceName(manager),
-			},
-			Port: &routev1.RoutePort{
-				TargetPort: intstr.FromInt(GatewayPort),
-			},
-			TLS: &routev1.TLSConfig{
-				InsecureEdgeTerminationPolicy: routev1.InsecureEdgeTerminationPolicyRedirect,
-				Termination:                   routev1.TLSTerminationEdge,
-			},
-		},
-	}
-}
-
-func getIngressSpec(manager *v1alpha1.CheManager) *v1beta1.Ingress {
-	pathType := v1beta1.PathTypeImplementationSpecific
-	return &v1beta1.Ingress{
-		ObjectMeta: metav1.ObjectMeta{
-			Name:      manager.Name,
-			Namespace: manager.Namespace,
-			Labels:    defaults.GetLabelsForComponent(manager, "external-access"),
-			Annotations: map[string]string{
-				"kubernetes.io/ingress.class":                       "nginx",
-				"nginx.ingress.kubernetes.io/proxy-read-timeout":    "3600",
-				"nginx.ingress.kubernetes.io/proxy-connect-timeout": "3600",
-				// "nginx.ingress.kubernetes.io/ssl-redirect":          "true", - do we need this?
-			},
-		},
-		Spec: v1beta1.IngressSpec{
-			Rules: []v1beta1.IngressRule{
-				{
-					Host: manager.Spec.Host,
-					IngressRuleValue: v1beta1.IngressRuleValue{
-						HTTP: &v1beta1.HTTPIngressRuleValue{
-							Paths: []v1beta1.HTTPIngressPath{
-								{
-									Path:     "/",
-									PathType: &pathType,
-									Backend: v1beta1.IngressBackend{
-										ServiceName: GetGatewayServiceName(manager),
-										ServicePort: intstr.FromInt(GatewayPort),
-									},
-								},
-							},
-						},
-					},
-				},
-			},
-		},
-	}
 }
